Ticket Histories
Any property change to a ticket should now create a new ticket history instance, which are listed on the details page. Applies to adding comments and attachments. Since there is no editing/deleting of comments or attachments, those obviously do not create a new history instance and are therfore not shown in the view.

Note that the tickets are not sorting correctly by date in index view. Eg, 10/10/2017 is considered lower sort order than 10/3/2017 when attempting to to sort dates ascending (older to newer).