Project Access & Profile
Projects Access
Inconsistent authentication conditions for devs and submitters
This is now based on the project helper method that finds if the user is on a project. instead of previously finding if the submitter or developer owned or were assigned to tickets for the given project. Pms are now redirected to index if they try to view the edit or assign views, or if they somehow post to those actions, if they are not on the specified project. Admins retain these accesses. Pms also cannot view a proejct details page if they are not on it. However, Pms still see the ticket title as a link instead of generic text. Options, build up the base view model or separate the assigned/all projects so that all list only contains projects they are not on.
Profile
Add profile picture and flesh out the profile page
Can upload from the manage account page, disposes if user currently has an image, image file, url, and created date are stored in ApplicationUser class as distinct properties; Could add editing features such as cropping before uploading	Profile page is set up with generic information